requirements.txt
pytest-postgresql
pytest-docker
confluent-kafka
psycopg[binary]
python-dotenv
